Moved to correct package, comments in english, detailed comment before class.
